refactor(task-session-manager): extract status-utils + idle-reconciliation
Split two more subsystems out of index.ts (1262 → 1088 lines):
- status-utils.ts: pure session-status helpers (extractTaskSummary,
  isActiveStatus, updateBackgroundJobFromOutput, normalizeLateCancelledTaskOutput,
  isLateCancelledTaskError, formatCancelledTaskStatusOutput).
- idle-reconciliation.ts: createIdleReconciler owning the two idle-timer Maps
  with schedule/clear methods; resolves circular deps via forward-ref pattern.

Also applies P1 Oracle finding: observeChatMessage now calls
continuationTokens.clearContinuation(sessionID) instead of the manual
invalidateContinuation + consumed.delete pair.

Pure extraction; behavior and exports unchanged. Full suite (1638 tests) green.
